Separate hardhat tasks into its own file
