+++ smart-contracts/tsconfig.json
@@ -7,5 +7,5 @@
     "outDir": "dist"
   },
   "include": ["./scripts", "./test"],
-  "files": ["./hardhat.config.ts", "./hardhat-utils.ts"]
+  "files": ["./hardhat.config.ts", "./hardhat-tasks.ts", "./hardhat-utils.ts"]
 }
